WindowsTaskSlaveRestarter deprecated and replaced by WindowsTaskAgentRestarter
