+++ src/main/java/de/tracetronic/jenkins/plugins/ecutest/extension/slave/WindowsTaskSlaveRestarter.java
@@ -5,11 +5,6 @@
  */
 package de.tracetronic.jenkins.plugins.ecutest.extension.slave;
 
-import hudson.Extension;
-import jenkins.slaves.restarter.SlaveRestarter;
-
-import java.io.IOException;
-
 /**
  * JNLP agent restarter based on Windows Task Scheduler.
  * <p>
@@ -18,74 +13,14 @@ import java.io.IOException;
  * </p>
  * <p>
  * In order to work a new task in the Windows Task Scheduler has to be created named <br>
- * <i>RESTART_JENKINS_AGENT</i> or renamed individually by system property <i>ecutest.taskName</i>.
+ * <i>RESTART_JENKINS_SLAVE</i> or renamed individually by system property <i>ecutest.taskName</i>.
  * This task should be configured with actions how to restart the agent.
  * </p>
  *
  * @since 1.8
  */
-@Extension
-public class WindowsTaskSlaveRestarter extends SlaveRestarter {
+@Deprecated
+public class WindowsTaskSlaveRestarter extends WindowsTaskAgentRestarter {
 
-    private static final long serialVersionUID = 1L;
 
-    /**
-     * Change the task name by invoking -Decutest.taskName={@code taskName} to Jenkins slave JVM
-     * or setting system property ecutest.taskName directly.
-     */
-    private static final String TASKNAME = System.getProperty("ecutest.taskName", "RESTART_JENKINS_SLAVE");
-
-    @Override
-    public boolean canWork() {
-        try {
-            final int ret = queryTask();
-            return ret == 0;
-        } catch (InterruptedException | IOException ignored) {
-            // no-op
-        }
-        return false;
-    }
-
-    @Override
-    public void restart() throws Exception {
-        final int ret = execTask();
-        throw new IOException("Failed restarting slave!\n"
-            + "Task completed with exit value: " + ret);
-    }
-
-    /**
-     * Queries the task scheduler.
-     *
-     * @return the process exit value
-     * @throws InterruptedException the interrupted exception
-     * @throws IOException          signals that an I/O exception has occurred
-     */
-    private int queryTask() throws InterruptedException, IOException {
-        return runProcess("/query");
-    }
-
-    /**
-     * Executes the task scheduler.
-     *
-     * @return the process exit value
-     * @throws InterruptedException the interrupted exception
-     * @throws IOException          signals that an I/O exception has occurred
-     */
-    private int execTask() throws InterruptedException, IOException {
-        return runProcess("/run");
-    }
-
-    /**
-     * Runs the task process with appropriate arguments.
-     *
-     * @param option the option
-     * @return the process exit value
-     * @throws IOException          signals that an I/O exception has occurred
-     * @throws InterruptedException the interrupted exception
-     */
-    private int runProcess(final String option) throws IOException, InterruptedException {
-        final ProcessBuilder procBuilder = new ProcessBuilder("schtasks.exe", option, "/tn", TASKNAME);
-        final Process proc = procBuilder.start();
-        return proc.waitFor();
-    }
 }
